do everything script file
This script stands on the jqGrid script, which I will upload later. Also, this is a little unwieldy. I will be making it more functional with namespaces, very soon. *Note - This was designed to dynamically update content served in Flash. I am leaving this version as is, because there might still be some use for it somewhere. It is simple to convert to HTML Article or Post.